"Added a search form container and input field placeholder in Home page"
The diff shows the addition of a search form with a container and input field in the Home page. The form has a placeholder text for the input field. However, the search button is present but not functional as its onClick event is not defined. The search form is designed to be responsive, adjusting its width based on the screen size.